+++ srcs/ft_lstsize.c
@@ -12,9 +12,9 @@
 
 #include <libft.h>
 
-int	ft_lstsize(t_list *lst)
+size_t	ft_lstsize(t_list *lst)
 {
-	int		i;
+	size_t		i;
 	t_list	*node;
 
 	i = 1;
