+++ srcs/ft_memchr.c
@@ -10,9 +10,11 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memchr(const void *src, int c, unsigned int n)
+#include <libft.h>
+
+void	*ft_memchr(const void *src, int c, size_t n)
 {
-	unsigned int	i;
+	size_t			i;
 	unsigned char	*res;
 
 	i = 0;
