+++ srcs/ft_striteri.c
@@ -10,7 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-void	ft_striteri(char *s, void (*f)(unsigned int, char*))
+#include <libft.h>
+
+void	ft_striteri(char *s, void (*f)(size_t, char*))
 {
 	int	i;
